requirements.txt
fastapi==0.100.0  # or newer
pydantic>=2.8,<3.0
python-dateutil==2.8.2
fetchai==0.1.41
python-dotenv==1.0.1
google.generativeai
flask
requests
deepgram-sdk
flask_cors
uagents==0.21.0
uvicorn>=0.30.1,<0.31.0
langchain==0.1.0
chromadb==0.4.22
openai==1.12.0
python-jose==3.3.0
passlib==1.7.4
PyJWT==2.8.0